started to clean up code and delete user is now woriking so that the usercan only delete themselves

+++ src/main/java/com/lina/individuelluppgift/user/UserController.java
@@ -1,6 +1,7 @@
 package com.lina.individuelluppgift.user;
 
 import com.lina.individuelluppgift.security.JwtService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,11 +18,14 @@ import java.util.Map;
 public class UserController {
 
     private final UserService userService;
-
+    private final JwtService jwtService;
+    private final UserRepository userRepository;
 
     @Autowired
-    public UserController(UserService userService) {
+    public UserController(UserService userService, JwtService jwtService, UserRepository userRepository) {
         this.userService = userService;
+        this.jwtService = jwtService;
+        this.userRepository = userRepository;
     }
 
 @Transactional
@@ -33,9 +37,19 @@ public class UserController {
 
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> deleteUser(@PathVariable Integer id){
+    public ResponseEntity<?> deleteUser(@PathVariable Integer id, HttpServletRequest request){
+
+        String token = request.getHeader("Authorization").substring(7);
+        String username = jwtService.extractUsername(token);
+        Integer currentUserId = userService.findUserIdByUsername(username);
+
+        if (!currentUserId.equals(id)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You are not authorized to delete this user");
+        }
+
         userService.deleteUser(id);
         return ResponseEntity.noContent().build();
     }
+
 
 }
